Adding updates to gui. Moved to postgres10. Using docker-compose to start it up.
Adding first pass at using celery. Works without docker (local python). Need to add to docker and get running.
