+++ docker-compose.yml
@@ -1,9 +1,66 @@
 version: '3'
+
 services:
-  magine:
-    build: .
-    command: python manage.py runserver 0.0.0.0:8000
+  db:
+    image: postgres:10
+    environment:
+    - POSTGRES_USER=postgres
+    - POSTGRES_PASSWORD=postgres
     volumes:
-      - ./_state/:/magine_home/_state/
+    - postgres_data:/var/lib/postgresql/data/
+    ports:
+    - 5432:5432
+    expose:
+    - "5432"
+    healthcheck:
+      test: ["CMD-SHELL", "pg_isready -U postgres"]
+      interval: 30s
+      timeout: 30s
+      retries: 6
+
+  magine:
+    restart: always
+    build: .
+    environment:
+    - POSTGRES_USER=postgres
+    - POSTGRES_PASSWORD=postgres
+    command: python manage.py runserver 0.0.0.0:8000
     ports:
       - "8000:8000"
+    depends_on:
+    - db
+    - redis
+
+  # Redis - result backend
+  redis:
+    image: redis:2.8.19
+    hostname: redis
+
+
+  # RabbitMQ - queue
+  rabbit:
+    hostname: rabbit
+    image: rabbitmq:3-management
+    environment:
+    - RABBITMQ_DEFAULT_USER=admin
+    - RABBITMQ_DEFAULT_PASS=mypass
+    ports:
+    - "5672:5672"
+    - "15672:15672"  # here, we can access rabbitmq management plugin
+
+  # Celery worker
+  worker:
+    build:
+      context: .
+      dockerfile: Dockerfile
+    command: sh ./run_celery.sh
+    volumes:
+    - .:/app
+    links:
+    - rabbit
+    - redis
+    depends_on:
+    - rabbit
+
+volumes:
+  postgres_data:
