+++ gui/views/enrichment.py
@@ -3,13 +3,11 @@ from django.shortcuts import render
 from django.template.loader import get_template
 from django.views import View
 
-import gui.forms as forms
+import gui.forms.enrichment as forms
+import gui.enrichment_functions.enrichr_helper  as enrichr_help
 
-from gui.data_functions.add_raptr_project import add_project_from_zip
 from.common import check_species_list
-from gui.enrichment_functions.enrichr_helper import return_table, \
-    model_to_json, return_table_from_model, add_enrichment
-from gui.models import Data, EnrichmentOutput
+from gui.models import EnrichmentOutput
 
 
 class EnrichmentResultsView(View):
@@ -19,7 +17,7 @@ class EnrichmentResultsView(View):
             genes = form.cleaned_data['list_of_species'].split(',')
             list_of_species = check_species_list(genes)
             ont = form.cleaned_data['enrichment']
-            data = return_table(list_of_species, ont)
+            data = enrichr_help.return_table(list_of_species, ont)
 
             template = get_template('simple_table_view.html', using='jinja2')
             return HttpResponse(template.render(data))
@@ -36,7 +34,8 @@ class ProjectEnrichmentView(View):
             project_name = form.cleaned_data['project_name']
             category = form.cleaned_data['category']
             dbs = form.cleaned_data['db']
-            data = return_table_from_model(project_name, category, dbs)
+            data = enrichr_help.return_table_from_model(project_name, category,
+                                                        dbs)
             template = get_template('simple_table_view.html', using='jinja2')
             return HttpResponse(template.render(data))
         else:
@@ -48,5 +47,5 @@ class ProjectEnrichmentView(View):
 
 def project_enrichment(request, project_name):
     ex = EnrichmentOutput.objects.filter(project_name=project_name).values()
-    data = model_to_json(ex)
+    data = enrichr_help.model_to_json(ex)
     return render(request, 'simple_table_view.html', data)
