+++ magine_gui_app/__init__.py
@@ -2,4 +2,5 @@ from __future__ import absolute_import
 
 # This will make sure the app is always imported when
 # Django starts so that shared_task will use this app.
-# from .celery import app as celery_app
+
+from .celery import app as celery_app
