+++ requirements.txt
@@ -1,7 +1,12 @@
-django==1.11.15
+Django>=1.11,<2.0
 jinja2
 django-picklefield
 django-import-export
+psycopg2-binary==2.7.4
+
+celery==4.2.1
+redis==2.10.6
+
 numpy>=1.9.0
 pandas>=0.23.0
 networkx>=2.1
@@ -15,10 +20,13 @@ pathos
 appdirs
 plotly==2.7
 seaborn
-py2cytoscape
-pydotplus
-IPython>=5.3.0
-jupyter==1.0.0
 wordcloud
 sortedcontainers
 statsmodels
+
+#py2cytoscape
+pydot
+pydotplus
+IPython>=5.3.0
+jupyter==1.0.0
+
