+++ run_server.sh
@@ -1,3 +1,4 @@
-export PYTHONPATH=~/PycharmProjects/Magine/:$PYTHONPATH
+source activate magine_36
+export PYTHONPATH=/run/media/pinojc/SharedData/PycharmProjects/PycharmProjects/Magine:$PYTHONPATH
 python manage.py runserver
